Extract methods from BusinessMatchingService to TradingPremisesService
