+++ app/controllers/businessmatching/updateservice/WhichCurrentTradingPremisesController.scala
@@ -27,7 +27,7 @@ import forms.{EmptyForm, Form2, InvalidForm, ValidForm}
 import models.businessmatching.updateservice.{TradingPremisesActivities => TradingPremisesForm}
 import models.businessmatching.{BusinessActivities, BusinessActivity}
 import models.tradingpremises.TradingPremises
-import services.businessmatching.BusinessMatchingService
+import services.businessmatching.{BusinessMatchingService, TradingPremisesService}
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.play.frontend.auth.AuthContext
 import uk.gov.hmrc.play.frontend.auth.connectors.AuthConnector
@@ -38,7 +38,8 @@ import scala.concurrent.Future
 
 class WhichCurrentTradingPremisesController @Inject()(val authConnector: AuthConnector,
                                                       val dataCacheConnector: DataCacheConnector,
-                                                      businessMatchingService: BusinessMatchingService) extends BaseController with RepeatingSection {
+                                                      businessMatchingService: BusinessMatchingService,
+                                                      tradingPremisesService: TradingPremisesService) extends BaseController with RepeatingSection {
 
   private val failure = InternalServerError("Could not get form data")
 
@@ -79,7 +80,7 @@ class WhichCurrentTradingPremisesController @Inject()(val authConnector: AuthCon
   private def updateTradingPremises(data: TradingPremisesForm, activity: BusinessActivity)
                                    (implicit ac: AuthContext, hc: HeaderCarrier): Future[_] =
     updateDataStrict[TradingPremises] { tradingPremises: Seq[TradingPremises] =>
-      businessMatchingService.addBusinessActivtiesToTradingPremises(data.index.toSeq, tradingPremises, activity, true)
+      tradingPremisesService.addBusinessActivtiesToTradingPremises(data.index.toSeq, tradingPremises, activity, true)
     }
 
   private def formData(implicit hc: HeaderCarrier, ac: AuthContext) = for {
