+++ app/controllers/businessmatching/updateservice/WhichTradingPremisesController.scala
@@ -29,7 +29,7 @@ import models.businessmatching.updateservice.{TradingPremisesActivities => Tradi
 import models.status.{NotCompleted, SubmissionReady}
 import models.tradingpremises.TradingPremises
 import services.StatusService
-import services.businessmatching.BusinessMatchingService
+import services.businessmatching.{BusinessMatchingService, TradingPremisesService}
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.play.frontend.auth.AuthContext
 import uk.gov.hmrc.play.frontend.auth.connectors.AuthConnector
@@ -43,7 +43,8 @@ class WhichTradingPremisesController @Inject()(
                                                 val authConnector: AuthConnector,
                                                 val dataCacheConnector: DataCacheConnector,
                                                 val statusService: StatusService,
-                                                val businessMatchingService: BusinessMatchingService
+                                                val businessMatchingService: BusinessMatchingService,
+                                                val tradingPremisesService: TradingPremisesService
                                               )() extends BaseController with RepeatingSection {
 
   def get(index: Int = 0) = Authorised.async {
@@ -109,7 +110,7 @@ class WhichTradingPremisesController @Inject()(
   private def updateTradingPremises(data: TradingPremises$, activity: BusinessActivity)
                                    (implicit ac: AuthContext, hc: HeaderCarrier): Future[_] =
     updateDataStrict[TradingPremises] { tradingPremises: Seq[TradingPremises] =>
-      businessMatchingService.addBusinessActivtiesToTradingPremises(data.index.toSeq, tradingPremises, activity, false)
+      tradingPremisesService.addBusinessActivtiesToTradingPremises(data.index.toSeq, tradingPremises, activity, false)
     }
 
   private def tradingPremises(implicit hc: HeaderCarrier, ac: AuthContext): Future[Seq[(TradingPremises, Int)]] =
