+++ app/services/businessmatching/BusinessMatchingService.scala
@@ -112,51 +112,4 @@ class BusinessMatchingService @Inject()(
 
   def activitiesToIterate(index: Int, activities: Set[BusinessActivity]) = activities.size > index + 1
 
-  def addBusinessActivtiesToTradingPremises(
-                                             indices: Seq[Int],
-                                             tradingPremises: Seq[TradingPremises],
-                                             activity: BusinessActivity,
-                                             remove: Boolean): Seq[TradingPremises] =
-    patchTradingPremisesBusinessActivities(tradingPremises){ (wdybd, index) =>
-      wdybd.copy({
-        if (indices contains index) {
-          wdybd.activities + activity
-        } else if (remove) {
-          wdybd.activities - activity
-        } else {
-          wdybd.activities
-        }
-      })
-    }
-
-  def removeBusinessActivitiesFromTradingPremises(
-                                                 tradingPremises: Seq[TradingPremises],
-                                                 existingActivities: Set[BusinessActivity],
-                                                 removeActivities: Set[BusinessActivity]): Seq[TradingPremises] =
-    patchTradingPremisesBusinessActivities(tradingPremises) { (wdybd, index) =>
-      wdybd.copy({
-        wdybd.activities diff removeActivities match {
-          case remainingActivities if remainingActivities.nonEmpty => remainingActivities
-          case _ => Set(existingActivities.head)
-        }
-      })
-    } map { tp =>
-      if(removeActivities contains MoneyServiceBusiness) {
-        tp.copy(msbServices = None)
-      } else {
-        tp
-      }
-    }
-
-
-  def patchTradingPremisesBusinessActivities(tradingPremises: Seq[TradingPremises])
-                                            (fn: ((WhatDoesYourBusinessDo, Int) => WhatDoesYourBusinessDo)): Seq[TradingPremises] =
-    tradingPremises.zipWithIndex map { case (tp, index) =>
-      tp.whatDoesYourBusinessDoAtThisAddress(
-        tp.whatDoesYourBusinessDoAtThisAddress.fold(WhatDoesYourBusinessDo(Set.empty)){ wdybd =>
-          fn(wdybd, index)
-        }
-      ).copy(hasAccepted = true)
-    }
-
 }
