+++ test/controllers/businessmatching/updateservice/UpdateServiceDateOfChangeControllerSpec.scala
@@ -42,7 +42,7 @@ import play.api.mvc.{Result, Results}
 import utils.{AuthorisedFixture, DependencyMocks, GenericTestHelper}
 import play.api.test.Helpers._
 import services.StatusService
-import services.businessmatching.BusinessMatchingService
+import services.businessmatching.{BusinessMatchingService, TradingPremisesService}
 import uk.gov.hmrc.http.cache.client.CacheMap
 import uk.gov.hmrc.play.frontend.auth.connectors.AuthConnector
 
@@ -60,22 +60,26 @@ class UpdateServiceDateOfChangeControllerSpec extends GenericTestHelper
 
     val request = addToken(authRequest)
 
+    val mockBusinessMatchingService = mock[BusinessMatchingService]
+    val mockTradingPremisesService = mock[TradingPremisesService]
+
     lazy val app = new GuiceApplicationBuilder()
       .disable[com.kenshoo.play.metrics.PlayModule]
       .overrides(bind[DataCacheConnector].to(mockCacheConnector))
       .overrides(bind[StatusService].to(mockStatusService))
       .overrides(bind[AuthConnector].to(self.authConnector))
-      .overrides(bind[BusinessMatchingService].to(mock[BusinessMatchingService]))
+      .overrides(bind[BusinessMatchingService].to(mockBusinessMatchingService))
+      .overrides(bind[TradingPremisesService].to(mockTradingPremisesService))
       .build()
 
     val controller = app.injector.instanceOf[UpdateServiceDateOfChangeController]
 
     when {
-      controller.businessMatchingService.updateModel(any())(any(),any(),any())
+      mockBusinessMatchingService.updateModel(any())(any(),any(),any())
     } thenReturn OptionT.some[Future, CacheMap](mockCacheMap)
 
     when {
-      controller.businessMatchingService.commitVariationData(any(),any(),any())
+      mockBusinessMatchingService.commitVariationData(any(),any(),any())
     } thenReturn OptionT.some[Future, CacheMap](mockCacheMap)
 
   }
@@ -105,7 +109,7 @@ class UpdateServiceDateOfChangeControllerSpec extends GenericTestHelper
           mockCacheSave[BusinessMatching]
 
           when {
-            controller.businessMatchingService.getModel(any(),any(),any())
+            mockBusinessMatchingService.getModel(any(),any(),any())
           } thenReturn OptionT.some[Future, BusinessMatching](BusinessMatching(
             activities = Some(BusinessActivities(Set(
               EstateAgentBusinessService,
@@ -202,7 +206,7 @@ class UpdateServiceDateOfChangeControllerSpec extends GenericTestHelper
       mockCacheSave[Msb]
 
       when {
-        controller.businessMatchingService.getModel(any(),any(),any())
+        mockBusinessMatchingService.getModel(any(),any(),any())
       } thenReturn OptionT.some[Future, BusinessMatching](BusinessMatching(
         activities = Some(BusinessActivities(Set(
           EstateAgentBusinessService,
@@ -212,7 +216,7 @@ class UpdateServiceDateOfChangeControllerSpec extends GenericTestHelper
       ))
 
       when {
-        controller.businessMatchingService.removeBusinessActivitiesFromTradingPremises(any(),any(),any())
+        mockTradingPremisesService.removeBusinessActivitiesFromTradingPremises(any(),any(),any())
       } thenReturn tradingPremises
 
       val result = controller.post("01/02/03/04/05/06")(request.withFormUrlEncodedBody(
@@ -223,7 +227,7 @@ class UpdateServiceDateOfChangeControllerSpec extends GenericTestHelper
 
       status(result) must be(SEE_OTHER)
 
-      verify(controller.businessMatchingService).updateModel(
+      verify(mockBusinessMatchingService).updateModel(
         eqTo(BusinessMatching(
           activities = Some(BusinessActivities(
             Set(TelephonePaymentService),
@@ -242,37 +246,37 @@ class UpdateServiceDateOfChangeControllerSpec extends GenericTestHelper
           hasAccepted = true
         )))(any(),any(),any())
 
-      verify(controller.dataCacheConnector).save[Seq[TradingPremises]](
+      verify(mockCacheConnector).save[Seq[TradingPremises]](
         eqTo(TradingPremises.key),
         eqTo(tradingPremises)
       )(any(),any(),any())
 
-      verify(controller.dataCacheConnector).save[Hvd](
+      verify(mockCacheConnector).save[Hvd](
         eqTo(Hvd.key),
         eqTo(None)
       )(any(),any(),any())
 
-      verify(controller.dataCacheConnector).save[Asp](
+      verify(mockCacheConnector).save[Asp](
         eqTo(Asp.key),
         eqTo(None)
       )(any(),any(),any())
 
-      verify(controller.dataCacheConnector).save[Supervision](
+      verify(mockCacheConnector).save[Supervision](
         eqTo(Supervision.key),
         eqTo(None)
       )(any(),any(),any())
 
-      verify(controller.dataCacheConnector).save[Msb](
+      verify(mockCacheConnector).save[Msb](
         eqTo(Msb.key),
         eqTo(None)
       )(any(),any(),any())
 
-      verify(controller.dataCacheConnector).save[Tcsp](
+      verify(mockCacheConnector).save[Tcsp](
         eqTo(Tcsp.key),
         eqTo(None)
       )(any(),any(),any())
 
-      verify(controller.dataCacheConnector).save[Eab](
+      verify(mockCacheConnector).save[Eab](
         eqTo(Eab.key),
         eqTo(None)
       )(any(),any(),any())
